requirements.txt
streamlit
langchain-groq
python-dotenv
textstat
nltk
streamlit-option-menu
